Fix ButtonLayout class; make it similar to CheckBoxLayout for consistency
Now buttons can be created in groups, unlike with the previous approach
where ButtonLayout accepted only a list of already existing Buttons.
Update function for now returns the activated button and further action
is done upon that information.

Maybe in future each button can get function that is supposed to be run
as a parameter, because for now to know which button was pressed
extra checks have to be done (such as string comparison).
This kind of defeates the purpose of having an abstraction over buttons.